extract most of the logic to a hook
